Calculate correct phase increment without improper rounding

+++ src/osc.h
@@ -30,7 +30,7 @@ private:
     uint16_t freq_hz_;
     uint8_t amplitude_;
 
-    uint32_t phase_;
+    double phase_;
     double phase_increment_;
 };
